# api/aether-app-gtwy-openapi3.yaml
# SPDX-FileCopyrightText: 2021-present Open Networking Foundation <info@opennetworking.org>
#
# SPDX-License-Identifier: Apache-2.0
info:
  contact:
    email: info@opennetworking.org
    name: Open Networking Foundation
    url: https://opennetworking.org
  description: "## Background\n\nThe Aether Application Gateway (AG) API provides a unified API for both control and metrics of the Aether platform.\nThe AG is device and application focused, enabling the application developer to implement intelligent control of Aether 4G/5G connectivity services, while abstracting some low-level intricacies of 4G/5G.\n"
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  title: Aether Application Gateway
  version: 1.0.0
openapi: 3.0.0

paths:
  /appgtwy/v1/{target}/enterprises/{enterprise-id}/sites/{site-id}/devices:
    get:
      operationId: getDevices
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App_Gtwy_Devices'
          description: GET OK 200
      summary: GET /enterprises/{enterprise-id}/site/{site-id}/devices
      tags:
        - ""
    parameters:
      - $ref: '#/components/parameters/target'
      - content:
          text/plain; charset=utf-8:
            schema:
              type: string
        description: key {enterprise-id}
        in: path
        name: enterprise-id
        required: true
      - content:
          text/plain; charset=utf-8:
            schema:
              type: string
        description: key {site-id}
        in: path
        name: site-id
        required: true
  /appgtwy/v1/{target}/enterprises/{enterprise-id}/sites/{site-id}/devices/{device-id}:
    get:
      operationId: getDevice
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App_Gtwy_Device'
          description: GET OK 200
      summary: GET /enterprises/{enterprise-id}/sites/{site-id}/devices/{device-id}
      tags:
        - ""
    parameters:
      - $ref: '#/components/parameters/target'
      - content:
          text/plain; charset=utf-8:
            schema:
              type: string
        description: key {enterprise-id}
        in: path
        name: enterprise-id
        required: true
      - content:
          text/plain; charset=utf-8:
            schema:
              type: string
        description: key {site-id}
        in: path
        name: site-id
        required: true
      - content:
          text/plain; charset=utf-8:
            schema:
              type: string
        description: key {device-id}
        in: path
        name: device-id
        required: true
components:
  parameters:
    target:
      content:
        text/plain; charset=utf-8:
          schema:
            type: string
      description: target (device in onos-config)
      in: path
      name: target
      required: true
  schemas:
    App_Gtwy_Devices:
#      additionalProperties:
#        $ref: '#/components/schemas/AdditionalPropertyTarget'
      description: The top level app gateway devices container
      properties:
        devices:
          description: List of app gateway devices
          items:
            $ref: '#/components/schemas/App_Gtwy_Device'
          title: ItemApp_Gtwy_Device
          type: array
          uniqueItems: true
          x-keys:
            - device-id
      title: App-Gtwy-Devices
      type: object
    App_Gtwy_Device:
      properties:
#        device:
#          description: Device Details
#          items:
#            $ref: '#/components/schemas/App_Gtwy_Device'
#          title: Application
#          type: array
#          uniqueItems: true
#          x-keys:
#            - application-id
#        connectivity-service:
#          description: The list for connectivity-service
#          items:
#            $ref: '#/components/schemas/Enterprises_Enterprise_Connectivity-service'
#          title: Connectivity-service
#          type: array
#          uniqueItems: true
#          x-keys:
#            - connectivity-service
        device-id:
          description: Link to device
          title: device-id
          type: string
#          x-leafref: ../../../device/device-id
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        imei: # int
          description: imei
          maxLength: 80
          minLength: 1
          title: imei
          type: string
        attached: # int
          description: attached
          maxLength: 80
          minLength: 1
          title: attached
          type: string
        ip:
          description: ip
          maxLength: 80
          minLength: 1
          title: ip
          type: string
        sim_iccid: # int
          description: attached
          maxLength: 80
          minLength: 1
          title: attached
          type: string
        device_groups:
          description: The list of device groups
          items:
            type: string
#          items:
#            $ref: '#/components/schemas/App_Gtwy_Device_Group'
          title: device-groups
          type: array
          uniqueItems: true
#          x-keys:
#            - connectivity-service
      required:
        - enterprise-id
      type: object
#    App_Gtwy_Device_Group:
#      properties:
#        device-group:
#          description: Link to device group
#          title: device-group
#          type: string
##          x-leafref: ../../../device-group/device-group-id
#        enable:
#          default: true
#          description: Enable this device group
#          title: enable
#          type: boolean
#      required:
#        - device-group
#      type: object